Class_Genius-main/class_Genius/requirements.txt
streamlit
pillow
pytesseract
fpdf
requests
beautifulsoup4